requirements.txt
Django==5.2.7
django-colorfield==0.14.0
django-import-export==2.8.0
pandas==2.3.3
plotly==6.5.1
python-dotenv==1.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
django-jquery==3.1.0
django-mathfilters==1.0.0
django-admin-logs==1.5.0
escpos==2.0.0
django-rangefilter==0.9.1
qrcode==8.2
python-barcode==0.16.1
pillow==12.1.0
openpyxl==3.1.5
